proa2/backend/requirements.txt
flask
twilio
python-telegram-bot
sqlite3
python-dotenv
